remove styled-components use less instead
We need it to be able to easily override styles.
Less instead of any other preprocessor was chosen to match superset one.

Similar to styled-components new css classes don't use css cascade and
have bblfsh prefix to be unique in a larger application.

Components themselves don't import styles to make it easier to integrate
to another codebase.

Signed-off-by: Maxim Sukharev <max@smacker.ru>